Interface and 2 implementations
